Refactor release tracking code into a separate helper file
